move preferences_schema.py to dialogs/preferences_dialog.py
