+++ napari/_qt/dialogs/preferences_dialog.py
